Extract bin lookup / network brand detection logic into separate class
COSDK-921
