+++ card/src/main/java/com/adyen/checkout/card/internal/data/api/DefaultDetectCardTypeRepository.kt
@@ -9,28 +9,21 @@
 package com.adyen.checkout.card.internal.data.api
 
 import com.adyen.checkout.card.internal.data.model.BinLookupCacheResult
-import com.adyen.checkout.card.internal.data.model.BinLookupRequest
-import com.adyen.checkout.card.internal.data.model.BinLookupResponse
-import com.adyen.checkout.card.internal.data.model.Brand
 import com.adyen.checkout.card.internal.data.model.DetectedCardType
 import com.adyen.checkout.core.common.AdyenLogLevel
 import com.adyen.checkout.core.common.CardBrand
 import com.adyen.checkout.core.common.internal.helper.adyenLog
 import com.adyen.checkout.core.common.internal.helper.bufferedChannel
-import com.adyen.checkout.core.common.internal.helper.runSuspendCatching
-import com.adyen.checkout.cse.internal.BaseCardEncryptor
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.receiveAsFlow
 import kotlinx.coroutines.launch
-import java.util.UUID
 
 internal class DefaultDetectCardTypeRepository(
-    private val cardEncryptor: BaseCardEncryptor,
-    private val binLookupService: BinLookupService,
     private val binLookupCache: BinLookupCache,
     private val localCardBrandDetectionService: LocalCardBrandDetectionService,
+    private val networkCardBrandDetectionService: NetworkCardBrandDetectionService,
 ) : DetectCardTypeRepository {
 
     private val _detectedCardTypesFlow: Channel<List<DetectedCardType>> = bufferedChannel()
@@ -43,7 +36,7 @@ internal class DefaultDetectCardTypeRepository(
         supportedCardBrands: List<CardBrand>,
         clientKey: String,
         coroutineScope: CoroutineScope,
-        type: String?
+        paymentMethodType: String?
     ) {
         adyenLog(AdyenLogLevel.DEBUG) { "detectCardType" }
         if (shouldFetchReliableTypes(cardNumber)) {
@@ -66,7 +59,7 @@ internal class DefaultDetectCardTypeRepository(
                         supportedCardBrands,
                         clientKey,
                         coroutineScope,
-                        type,
+                        paymentMethodType,
                     )
                 }
             }
@@ -82,7 +75,7 @@ internal class DefaultDetectCardTypeRepository(
         supportedCardBrands: List<CardBrand>,
         clientKey: String,
         coroutineScope: CoroutineScope,
-        type: String?
+        paymentMethodType: String?
     ) {
         if (publicKey != null) {
             adyenLog(AdyenLogLevel.DEBUG) { "Launching Bin Lookup" }
@@ -94,7 +87,7 @@ internal class DefaultDetectCardTypeRepository(
                     publicKey,
                     supportedCardBrands,
                     clientKey,
-                    type,
+                    paymentMethodType,
                 )?.let {
                     _detectedCardTypesFlow.send(it)
                 }
@@ -116,71 +109,29 @@ internal class DefaultDetectCardTypeRepository(
         publicKey: String,
         supportedCardBrands: List<CardBrand>,
         clientKey: String,
-        type: String?
+        paymentMethodType: String?
     ): List<DetectedCardType>? {
         val bin = getBin(cardNumber)
         binLookupCache.setFetching(bin)
 
-        val binLookupResponse = makeBinLookup(cardNumber, publicKey, supportedCardBrands, clientKey, type)
-        return if (binLookupResponse == null) {
+        val detectedCardTypes = networkCardBrandDetectionService.getCardBrands(
+            cardNumber,
+            publicKey,
+            supportedCardBrands,
+            clientKey,
+            paymentMethodType,
+        )
+        if (detectedCardTypes == null) {
             binLookupCache.remove(bin)
-            null
+            return null
         } else {
-            val detectedCardTypes = mapResponse(binLookupResponse)
             binLookupCache.setCachedResults(bin, detectedCardTypes)
-            detectedCardTypes
+            return detectedCardTypes
         }
     }
 
     private fun getBin(cardNumber: String): String {
         return cardNumber.take(REQUIRED_BIN_SIZE)
-    }
-
-    private suspend fun makeBinLookup(
-        cardNumber: String,
-        publicKey: String,
-        supportedCardBrands: List<CardBrand>,
-        clientKey: String,
-        type: String?
-    ): BinLookupResponse? {
-        return runSuspendCatching {
-            val encryptedBin = cardEncryptor.encryptBin(cardNumber, publicKey)
-            val cardBrands = supportedCardBrands.map { it.txVariant }
-            val request = BinLookupRequest(encryptedBin, UUID.randomUUID().toString(), cardBrands, type)
-
-            binLookupService.makeBinLookup(
-                request = request,
-                clientKey = clientKey,
-            )
-        }
-            .onFailure { e -> adyenLog(AdyenLogLevel.ERROR, e) { "checkCardType - Failed to do bin lookup" } }
-            .getOrNull()
-    }
-
-    private fun mapResponse(binLookupResponse: BinLookupResponse): List<DetectedCardType> {
-        adyenLog(AdyenLogLevel.DEBUG) { "handleBinLookupResponse" }
-        adyenLog(AdyenLogLevel.VERBOSE) { "Brands: ${binLookupResponse.brands}" }
-
-        // Any null or unmapped values are ignored, a null response becomes an empty list
-        return binLookupResponse.brands.orEmpty().mapNotNull { brandResponse ->
-            if (brandResponse.brand == null) return@mapNotNull null
-            val cardBrand = CardBrand(txVariant = brandResponse.brand)
-            DetectedCardType(
-                cardBrand = cardBrand,
-                isReliable = true,
-                enableLuhnCheck = brandResponse.enableLuhnCheck == true,
-                cvcPolicy = Brand.FieldPolicy.parse(
-                    brandResponse.cvcPolicy ?: Brand.FieldPolicy.REQUIRED.value,
-                ),
-                expiryDatePolicy = Brand.FieldPolicy.parse(
-                    brandResponse.expiryDatePolicy ?: Brand.FieldPolicy.REQUIRED.value,
-                ),
-                isSupported = brandResponse.supported != false,
-                panLength = brandResponse.panLength,
-                paymentMethodVariant = brandResponse.paymentMethodVariant,
-                localizedBrand = brandResponse.localizedBrand,
-            )
-        }
     }
 
     companion object {
