+++ card/src/main/java/com/adyen/checkout/card/internal/data/api/DetectCardTypeRepository.kt
@@ -24,6 +24,6 @@ internal interface DetectCardTypeRepository {
         supportedCardBrands: List<CardBrand>,
         clientKey: String,
         coroutineScope: CoroutineScope,
-        type: String? = null
+        paymentMethodType: String? = null
     )
 }
